chore(json-patch): move domain service test to core package and add author
Why: The test for JsonPatchDomainService was placed in the infra test
package but the class under test lives in core; moving it aligns the test
package with the production package. JsonMergePatchService was also missing
the standard @author tag present on its sibling interfaces.
